Implementation: (StreamHandler) implementation of the class for managing input and output

+++ compression-decompression/IStreamInterface.h
@@ -6,7 +6,7 @@
 class IStreamInterface
 {
 public:
-	virtual void readData(std::vector<char>& buffer, long long size) = 0;
+	virtual void readData(std::vector<char>& buffer, int size) = 0;
 	virtual void readData(int& size) = 0;
 	virtual void writeData(const std::vector<char>& buffer) = 0;
 	virtual void writeData(int& size) = 0;
